Split up bqplot viewer code into sub-directories
