--- glue_jupyter/bqplot/__init__.py
@@ -1,1 +1,0 @@
-from .view import BqplotScatterView, BqplotImageView, BqplotHistogramView
