+++ glue_jupyter/bqplot/image/__init__.py
@@ -1,0 +1,1 @@
+from .viewer import BqplotImageView
