+++ glue_jupyter/bqplot/scatter/__init__.py
@@ -1,0 +1,1 @@
+from .viewer import BqplotScatterView
