+++ glue_jupyter/bqplot/scatter/layer_artist.py
@@ -10,8 +10,8 @@ from glue.viewers.scatter.state import ScatterLayerState
 from glue.core.exceptions import IncompatibleAttribute
 from glue.viewers.common.layer_artist import LayerArtist
 
-from ..link import link, dlink, calculation, link_component_id_to_select_widget, on_change
-from ..utils import colormap_to_hexlist, debounced, float_or_none
+from ...link import link, dlink, calculation, link_component_id_to_select_widget, on_change
+from ...utils import colormap_to_hexlist, debounced, float_or_none
 import glue_jupyter.widgets
 from glue.viewers.matplotlib.state import (MatplotlibDataViewerState,
                                            MatplotlibLayerState,
