+++ glue_jupyter/bqplot/view.py
@@ -10,10 +10,11 @@ from glue.core.command import ApplySubsetState
 
 import glue_jupyter.widgets.component
 from ..view import IPyWidgetView
-from ..link import link, dlink, calculation, link_component_id_to_select_widget, on_change
+from ..link import link, on_change
 from ..utils import float_or_none
 
 ICON_WIDTH = 20
+
 icon_brush = widgets.Image.from_file(glue.icons.icon_path('glue_square', icon_format='svg'), width=ICON_WIDTH)
 icon_pan = widgets.Image.from_file(glue.icons.icon_path('glue_move', icon_format='svg'), width=ICON_WIDTH)
 icon_brush_x = widgets.Image.from_file(glue.icons.icon_path('glue_xrange_select', icon_format='svg'), width=ICON_WIDTH)
@@ -26,9 +27,6 @@ class BqplotBaseView(IPyWidgetView):
     allow_duplicate_data = False
     allow_duplicate_subset = False
     is2d = True
-    # _state_cls = ImageViewerState
-    # _data_artist_cls = BqplotImageLayerArtist
-    # _subset_artist_cls = BqplotImageLayerArtist
 
     def __init__(self, session, state=None):
         super(BqplotBaseView, self).__init__(session, state=state)
@@ -175,21 +173,6 @@ class BqplotBaseView(IPyWidgetView):
                                        subset_state=subset_state,
                                        use_current=use_current)
                 self._session.command_stack.do(cmd)
-        # else:
-        #     # Make sure we force a redraw to get rid of the ROI
-        #    self.axes.figure.canvas.draw()
-
-    def apply_roi(self, roi, use_current=False):
-        with self.output_widget:
-            if len(self.layers) > 0:
-                subset_state = self._roi_to_subset_state(roi)
-                cmd = ApplySubsetState(data_collection=self._data,
-                                       subset_state=subset_state,
-                                       use_current=use_current)
-                self._session.command_stack.do(cmd)
-            else:
-                # Make sure we force a redraw to get rid of the ROI
-                self.axes.figure.canvas.draw()
 
     def _roi_to_subset_state(self, roi):
         # TODO: copy paste from glue/viewers/image/qt/data_viewer.py#L66
@@ -216,27 +199,8 @@ class BqplotBaseView(IPyWidgetView):
             self.scale_y.min = float(self.state.y_min)
             self.scale_y.max = float(self.state.y_max)
 
-    def get_data_layer_artist(self, layer=None, layer_state=None):
-        if layer.ndim == 1:
-            if self.is2d:
-                cls = BqplotScatterLayerArtist
-            else:
-                cls = BqplotHistogramLayerArtist
-        else:
-            cls = BqplotImageLayerArtist
-        layer = self.get_layer_artist(cls, layer=layer, layer_state=layer_state)
-        self._add_layer_tab(layer)
-        return layer
-
-    def get_subset_layer_artist(self, layer=None, layer_state=None):
-        if layer.ndim == 1:
-            if self.is2d:
-                cls = BqplotScatterLayerArtist
-            else:
-                cls = BqplotHistogramLayerArtist
-        else:
-            cls = BqplotImageLayerArtist
-        layer = self.get_layer_artist(cls, layer=layer, layer_state=layer_state)
+    def get_layer_artist(self, cls, layer=None, layer_state=None):
+        layer = super().get_layer_artist(cls, layer=layer, layer_state=layer_state)
         self._add_layer_tab(layer)
         return layer
 
@@ -246,148 +210,4 @@ class BqplotBaseView(IPyWidgetView):
         self.last_msg = message
 
     def redraw(self):
-        pass # print('redraw view', self.state.x_att, self.state.y_att)
-
-    def redraw(self):
         pass
-
-
-
-from glue.viewers.image.state import ImageViewerState
-from glue.viewers.scatter.state import ScatterViewerState
-from glue.viewers.image.composite_array import CompositeArray
-
-from .image import BqplotImageLayerArtist
-from .scatter import BqplotScatterLayerArtist
-from ..utils import rgba_to_png_data
-import numpy as np
-
-class BqplotImageView(BqplotBaseView):
-    allow_duplicate_data = False
-    allow_duplicate_subset = False
-    _state_cls = ImageViewerState
-    large_data_size = 2e7
-
-    def __init__(self, session):
-        super(BqplotImageView, self).__init__(session)
-        on_change([(self.state, 'aspect')])(self._sync_figure_aspect)
-        self._sync_figure_aspect()
-
-    def _sync_figure_aspect(self):
-        with self.figure.hold_trait_notifications():
-            if self.state.aspect == 'equal':
-                self.figure.max_aspect_ratio = 1
-                self.figure.min_aspect_ratio = 1
-            else:
-                self.figure.min_aspect_ratio = bqplot.Figure.min_aspect_ratio.default_value
-                self.figure.max_aspect_ratio = bqplot.Figure.max_aspect_ratio.default_value
-
-    def get_data_layer_artist(self, layer=None, layer_state=None):
-        if layer.ndim == 1:
-            cls = BqplotScatterLayerArtist
-        else:
-            cls = BqplotImageLayerArtist
-        layer = self.get_layer_artist(cls, layer=layer, layer_state=layer_state)
-        self._add_layer_tab(layer)
-        return layer
-
-    def get_subset_layer_artist(self, layer=None, layer_state=None):
-        if layer.ndim == 1:
-            cls = BqplotScatterLayerArtist
-        else:
-            cls = BqplotImageLayerArtist
-        layer = self.get_layer_artist(cls, layer=layer, layer_state=layer_state)
-        self._add_layer_tab(layer)
-        return layer
-
-    def create_tab(self):
-        super(BqplotImageView, self).create_tab()
-        self.widgets_aspect = widgets.Checkbox(description='Equal aspect ratio')
-        aspect_mapping = {'equal': True, 'auto': False}
-        aspect_mapping_inverse = {True: 'equal', False: 'auto'}
-        link((self.state, 'aspect'), (self.widgets_aspect, 'value'), lambda x: aspect_mapping[x], lambda x: aspect_mapping_inverse[x])
-
-        self.tab_general.children += (self.widgets_aspect,)
-
-
-
-class BqplotScatterView(BqplotBaseView):
-
-    allow_duplicate_data = False
-    allow_duplicate_subset = False
-    _state_cls = ScatterViewerState
-    _data_artist_cls = BqplotScatterLayerArtist
-    _subset_artist_cls = BqplotScatterLayerArtist
-    large_data_size = 1e7
-
-    def create_tab(self):
-        super(BqplotScatterView, self).create_tab()
-        self.widgets_axis = []
-        for i, axis_name in enumerate('xy'):
-            if hasattr(self.state, axis_name + '_att_helper'):
-                widget_axis = glue_jupyter.widgets.component.Component(
-                    self.state, axis_name + '_att', label=axis_name + ' axis'
-                )
-                self.widgets_axis.append(widget_axis)
-        self.tab_general.children += tuple(self.widgets_axis)
-
-
-from .histogram import BqplotHistogramLayerArtist, HistogramViewerState
-
-class BqplotHistogramView(BqplotBaseView):
-
-    allow_duplicate_data = False
-    allow_duplicate_subset = False
-    _state_cls = HistogramViewerState
-    _data_artist_cls = BqplotHistogramLayerArtist
-    _subset_artist_cls = BqplotHistogramLayerArtist
-    large_data_size = 1e5
-    is2d = False
-
-    def create_tab(self):
-        super(BqplotHistogramView, self).create_tab()
-        self.button_normalize = widgets.ToggleButton(
-            value=False, description='normalize', tooltip='Normalize histogram')
-        link((self.button_normalize, 'value'), (self.state, 'normalize'))
-
-        self.button_cumulative = widgets.ToggleButton(
-            value=False, description='cumulative', tooltip='cumulative histogram')
-        link((self.button_cumulative, 'value'), (self.state, 'cumulative'))
-
-
-        self.widgets_axis = []
-        for i, axis_name in enumerate('x'):
-            if hasattr(self.state, axis_name + '_att_helper'):
-                widget_axis = glue_jupyter.widgets.component.Component(
-                    self.state, axis_name + '_att', label=axis_name + ' axis'
-                )
-                self.widgets_axis.append(widget_axis)
-        # @on_change([(self.state, 'hist_n_bin')])
-        # def trigger():
-
-        # self.widget_hist_x_min = widgets.FloatText(description='x min')
-        # link((self.state, 'hist_x_min'), (self.widget_hist_x_min, 'value'))
-
-        # self.widget_hist_n_bin = widgets.IntSlider(min=1, max=1000, step=1, description='bins')
-        # link((self.state, 'hist_n_bin'), (self.widget_hist_n_bin, 'value'))
-
-        self.tab_general.children += (widgets.HBox([self.button_normalize, self.button_cumulative] ),) + tuple(self.widgets_axis)
-
-
-    def _roi_to_subset_state(self, roi):
-        # TODO: copy paste from glue/viewers/histogram/qt/data_viewer.py
-        # TODO Does subset get applied to all data or just visible data?
-
-        bins = self.state.bins
-
-        x = roi.to_polygon()[0]
-        lo, hi = min(x), max(x)
-
-        if lo >= bins.min():
-            lo = bins[bins <= lo].max()
-        if hi <= bins.max():
-            hi = bins[bins >= hi].min()
-
-        roi_new = RangeROI(min=lo, max=hi, orientation='x')
-
-        return roi_to_subset_state(roi_new, x_att=self.state.x_att)
